+++ packages/react-composites/src/composites/CallComposite/selectors/baseSelectors.ts
@@ -2,8 +2,8 @@
 // Licensed under the MIT license.
 
 import { CallState as SDKCallStatus } from '@azure/communication-calling';
-import { toFlatCommunicationIdentifier } from 'acs-ui-common';
-import { CallState, DeviceManagerState } from 'calling-stateful-client';
+import { toFlatCommunicationIdentifier } from '@internal/acs-ui-common';
+import { CallState, DeviceManagerState } from '@internal/calling-stateful-client';
 import { CallAdapterState, CallCompositePage } from '../adapter/CallAdapter';
 
 export const getCall = (state: CallAdapterState): CallState | undefined => state.call;