+++ packages/react-composites/src/composites/ChatComposite/ChatScreen.tsx
@@ -11,7 +11,7 @@ import {
   ParticipantList,
   SendBox,
   TypingIndicator
-} from 'react-components';
+} from '@internal/react-components';
 import { useAdapter } from './adapter/ChatAdapterProvider';
 import { useAdaptedSelector } from './hooks/useAdaptedSelector';
 import { usePropsFor } from './hooks/usePropsFor';
@@ -26,6 +26,7 @@ import {
   participantListStack,
   participantListStyle
 } from './styles/Chat.styles';
+import { CHAT_UI_IDS } from './identifiers';
 
 export type ChatScreenProps = {
   showParticipantPane?: boolean;
@@ -71,12 +72,13 @@ export const ChatScreen = (props: ChatScreenProps): JSX.Element => {
   );
 
   return (
-    <Stack id="chat-composite" className={chatContainer} grow>
+    <Stack data-ui-id={CHAT_UI_IDS.chatScreen} className={chatContainer} grow>
       <ChatHeader {...headerProps} />
       <Stack className={chatArea} horizontal grow>
         <Stack className={chatWrapper} grow>
           <ThreadStatus {...threadStatusProps} />
           <MessageThread
+            data-ui-id={CHAT_UI_IDS.messageThread}
             {...messageThreadProps}
             onRenderAvatar={onRenderMessageAvatar}
             onRenderMessage={onRenderMessage}