+++ packages/react-composites/src/composites/ChatComposite/hooks/useHandlers.ts
@@ -1,8 +1,8 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { CommonProperties } from 'acs-ui-common';
-import { DefaultChatHandlers } from 'chat-component-bindings';
+import { CommonProperties } from '@internal/acs-ui-common';
+import { DefaultChatHandlers } from '@internal/chat-component-bindings';
 
 import { ReactElement } from 'react';
 import memoizeOne from 'memoize-one';