+++ packages/react-composites/CHANGELOG.md
@@ -1,4 +1,4 @@
-# Change Log - react-composites
+# Change Log - @internal/react-composites
 
 This log was last generated on Fri, 09 Jul 2021 20:41:33 GMT and should not be manually modified.
 
@@ -80,7 +80,7 @@ Fri, 21 May 2021 16:16:28 GMT
 - Introduce common identifier format ([PR #315](https://github.com/azure/communication-ui-library/pull/315) by prprabhu@microsoft.com)
 - Rename DeviceManager to avoid name conflict ([PR #319](https://github.com/azure/communication-ui-library/pull/319) by prprabhu@microsoft.com)
 - removed all Text/Label children of VideoTile, also renamed local/remoteParticipantName to local/remoteParticipantDisplay name and removed label props from RemoteVideoTile  ([PR #313](https://github.com/azure/communication-ui-library/pull/313) by alcail@microsoft.com)
-- Move common type to acs-ui-common ([PR #303](https://github.com/azure/communication-ui-library/pull/303) by prprabhu@microsoft.com)
+- Move common type to @internal/acs-ui-common ([PR #303](https://github.com/azure/communication-ui-library/pull/303) by prprabhu@microsoft.com)
 - add event implemeration to calling adapter ([PR #311](https://github.com/azure/communication-ui-library/pull/311) by jinan@microsoft.com)
 - [#2404092] Adding stateful attributes to CallClientProvider for starting a call with camera on/off ([PR #292](https://github.com/azure/communication-ui-library/pull/292) by anjulgarg@live.com)
 - Move legacy mappers, providers and hooks to 1to1 composite ([PR #309](https://github.com/azure/communication-ui-library/pull/309) by easony@microsoft.com)
@@ -99,7 +99,7 @@ Fri, 21 May 2021 16:16:28 GMT
 - Prune exports and add chat composite to @azure/communication-react ([PR #276](https://github.com/azure/communication-ui-library/pull/276) by mail@jamesburnside.com)
 - migrate call composite to statuful ([PR #290](https://github.com/azure/communication-ui-library/pull/290) by easony@microsoft.com)
 - added CallingTheme and used it for EndCallButton styles ([PR #249](https://github.com/azure/communication-ui-library/pull/249) by alcail@microsoft.com)
-- Create composite package. Migrate files from react-components to new package. ([PR #188](https://github.com/azure/communication-ui-library/pull/188) by mail@jamesburnside.com)
+- Create composite package. Migrate files from @internal/react-components to new package. ([PR #188](https://github.com/azure/communication-ui-library/pull/188) by mail@jamesburnside.com)
 - First stab at GroupCallAdapter interface ([PR #216](https://github.com/azure/communication-ui-library/pull/216) by domessin@microsoft.com)
 - rename declarative packages to stateful ([PR #250](https://github.com/azure/communication-ui-library/pull/250) by domessin@microsoft.com)
 - Fix copyright header to MIT and add LICENSE files ([PR #225](https://github.com/azure/communication-ui-library/pull/225) by domessin@microsoft.com)