+++ samples/Chat/CHANGELOG.md
@@ -19,4 +19,4 @@ Fri, 21 May 2021 16:16:28 GMT
 
 ### Changes
 
-- Bump react-composites to v1.0.0-beta.1 ([PR #359](https://github.com/azure/communication-ui-library/pull/359) by allenhwang@microsoft.com)
+- Bump @internal/react-composites to v1.0.0-beta.1 ([PR #359](https://github.com/azure/communication-ui-library/pull/359) by allenhwang@microsoft.com)